feat: make FakeAuthenticator avialable to require and used outside of tests
